refactor: Add react-multi-carousel and react-slick for carousel functionality
