+++ package.json
@@ -17,9 +17,13 @@
     "react-i18next": "^14.1.2",
     "react-icons": "^5.2.1",
     "react-modal": "3.16.1",
+    "react-multi-carousel": "^2.8.5",
     "react-router-dom": "^6.23.1",
     "react-scripts": "latest",
+    "react-slick": "^0.30.2",
     "resolve-url-loader": "latest",
+    "slick-carousel": "^1.8.1",
+    "swiper": "^11.1.4",
     "tailwindcss": "latest",
     "web-vitals": "latest"
   },
